+++ apps/nextjs/src/app/(dashboard)/dashboard/_components/sidebar/nav-main/components.tsx
@@ -1,6 +1,6 @@
 import type { LucideIcon } from "lucide-react";
 import Link from "next/link";
-import { BookOpen, Bot, ChevronRight, SquareTerminal } from "lucide-react";
+import { ChevronRight } from "lucide-react";
 
 import {
   Collapsible,
@@ -128,61 +128,3 @@ export function EmptyNavItems() {
     </SidebarGroup>
   );
 }
-
-export const dashboardNavItem: NavMainItem = {
-  title: "Dashboard",
-  url: "/dashboard",
-  icon: SquareTerminal,
-  isActive: true,
-  // items: [
-  //   {
-  //     title: "History",
-  //     url: "#",
-  //   },
-  // ],
-};
-
-export const navMainItems: NavMainItem[] = [
-  {
-    title: "Profiles",
-    url: "/dashboard/profiles",
-    icon: Bot,
-    items: [
-      {
-        title: "Genesis",
-        url: "#",
-      },
-      {
-        title: "Explorer",
-        url: "#",
-      },
-      {
-        title: "Quantum",
-        url: "#",
-      },
-    ],
-  },
-  {
-    title: "Teams",
-    url: "#",
-    icon: BookOpen,
-    items: [
-      {
-        title: "Introduction",
-        url: "#",
-      },
-      {
-        title: "Get Started",
-        url: "#",
-      },
-      {
-        title: "Tutorials",
-        url: "#",
-      },
-      {
-        title: "Changelog",
-        url: "#",
-      },
-    ],
-  },
-];
