+++ apps/nextjs/src/app/(dashboard)/dashboard/_components/sidebar/nav-main/nav-main.tsx
@@ -1,4 +1,4 @@
-import { Settings2 } from "lucide-react";
+import { BookOpen, Bot, Settings2, SquareTerminal } from "lucide-react";
 
 import {
   Collapsible,
@@ -11,22 +11,61 @@ import type { NavMainItem } from "./components";
 import { getMyOrganizations } from "~/app/server-actions/organizations/actions";
 import {
   CollapsibleMenuNavItem,
-  dashboardNavItem,
-  EmptyNavItems,
-  navMainItems,
   SidebarMenuButtonCollapsibleTrigger,
   SidebarMenuCollapsibleContent,
 } from "./components";
 
-export function NavMain() {
-  if (navMainItems.length === 0) {
-    return <EmptyNavItems />;
-  }
+export const navMainItems: NavMainItem[] = [
+  {
+    title: "Profiles",
+    url: "/dashboard/profiles",
+    icon: Bot,
+    items: [
+      {
+        title: "Genesis",
+        url: "#",
+      },
+      {
+        title: "Explorer",
+        url: "#",
+      },
+      {
+        title: "Quantum",
+        url: "#",
+      },
+    ],
+  },
+  {
+    title: "Teams",
+    url: "#",
+    icon: BookOpen,
+    items: [
+      {
+        title: "Introduction",
+        url: "#",
+      },
+      {
+        title: "Get Started",
+        url: "#",
+      },
+      {
+        title: "Tutorials",
+        url: "#",
+      },
+      {
+        title: "Changelog",
+        url: "#",
+      },
+    ],
+  },
+];
 
+export function NavMain() {
   return (
     <SidebarGroup>
       <SidebarMenu>
         <DashboardCollapsibleNavItem />
+
         <OrganizationsCollapsibleMenuNavItem />
 
         {navMainItems.map((item) => (
@@ -66,6 +105,18 @@ async function OrganizationsCollapsibleMenuNavItem() {
   );
 }
 
+export const dashboardNavItem: NavMainItem = {
+  title: "Dashboard",
+  url: "/dashboard",
+  icon: SquareTerminal,
+  isActive: true,
+  // items: [
+  //   {
+  //     title: "History",
+  //     url: "#",
+  //   },
+  // ],
+};
 function DashboardCollapsibleNavItem() {
   return (
     <Collapsible
